api/requirements.txt
argh==0.26.2
cloudinary==1.25.0
    certifi==2020.12.5
    six==1.15.0
    urllib3==1.26.4
django-bulk-update==2.2.0
    Django==3.2
        asgiref==3.3.4
        pytz==2021.1
        sqlparse==0.4.1
django-celery-beat==2.2.0
    celery==5.0.5
        billiard==3.6.4.0
        click==7.1.2
        click-didyoumean==0.0.3
            # click==7.1.2
        click-plugins==1.1.1
            # click==7.1.2
        click-repl==0.1.6
            # click==7.1.2
            prompt-toolkit==3.0.18
                wcwidth==0.2.5
            # six==1.15.0
        kombu==5.0.2
            amqp==5.0.6
                vine==5.0.0
        # pytz==2021.1
        # vine==5.0.0
    # Django==3.2
        # asgiref==3.3.4
        # pytz==2021.1
        # sqlparse==0.4.1
    django-timezone-field==4.1.2
        # Django==3.2
            # asgiref==3.3.4
            # pytz==2021.1
            # sqlparse==0.4.1
        # pytz==2021.1
    python-crontab==2.5.1
        python-dateutil==2.8.1
            # six==1.15.0
django-dynamic-db-router==0.3.0
    # Django==3.2
        # asgiref==3.3.4
        # pytz==2021.1
        # sqlparse==0.4.1
django-filter==2.4.0
    # Django==3.2
        # asgiref==3.3.4
        # pytz==2021.1
        # sqlparse==0.4.1
django-mail-factory==0.22
    html2text==2020.1.16
djangorestframework-bulk==0.2.1
    # Django==3.2
        # asgiref==3.3.4
        # pytz==2021.1
        # sqlparse==0.4.1
    djangorestframework==3.12.4
        # Django==3.2
            # asgiref==3.3.4
            # pytz==2021.1
            # sqlparse==0.4.1
    # setuptools
djangorestframework-simplejwt==4.6.0
    # Django==3.2
        # asgiref==3.3.4
        # pytz==2021.1
        # sqlparse==0.4.1
    # djangorestframework==3.12.4
        # Django==3.2
            # asgiref==3.3.4
            # pytz==2021.1
            # sqlparse==0.4.1
    PyJWT==2.1.0
gunicorn==20.1.0
    # setuptools
httptools==0.1.1
jwt==1.2.0
    cryptography==3.4.7
        cffi==1.14.5
            pycparser==2.20
Markdown==3.3.4
pipfreeze==2.0.3
psycopg2==2.8.6
python-slugify==5.0.0
    text-unidecode==1.3
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
    # certifi==2020.12.5
    chardet==4.0.0
    idna==2.10
    # urllib3==1.26.4
singleton-decorator==1.0.0
uvicorn==0.13.4
    # click==7.1.2
    h11==0.12.0
uvloop==0.15.2
watchdog==2.0.3
